chat_project/requirements.txt
Django==5.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.68.0
